requirements.txt
biopython>=1.84
matplotlib>=3.8.0
seaborn>=0.13.0